+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 ## 0.2.0.2 -- 2026-08-27
 
 * some domain logic functions implemented
+
+## 0.2.1.0 -- 2026-08-29
+
+* Change domain type name to Case, modul name to Worklog.Case
+* implement OnDesk, Deferred Day and Archived in Status
+* remove caseStatus from export list, API functions for Status are full